added check path bin converted before copying

+++ appV2/process_pcl.py
@@ -36,7 +36,10 @@ def convert_bin_to_las(path_to_data='/home/data/'):
     file_list = glob.glob(path_to_data + 'bin/*.bin')
     for file in file_list:
         opl.convertBin2LAS(file, deleteBin=True)
-        os.rename(file + '.las', path_to_data + 'las_raw/' + file.split('/')[-1][:-4] + '.las')
+        if os.path.isfile(file + '.las'):
+            os.rename(file + '.las', path_to_data + 'las_raw/' + file.split('/')[-1][:-4] + '.las')
+        else:
+            print(file, ' does not exist.')
 
 def rotate_point_clouds(extrinsic=[0,0,0,0,0,0], z_range='[-20:20]', crop_corners='([-20, 10], [-5, 5])', path_to_data='/home/data/'):
     """
